finalizing Routeable interface, organizing code, adding documentation
